Se corrigio la consulta del tipo inner join entre articulo y categoria, se agregaron las modificaciones den la clase articulo

+++ GestorVentas/Models/Almacen/Articulo/ArticuloVM.cs
@@ -17,7 +17,7 @@ namespace GestorVentas.Models.Almacen.Articulo
         
         public string Nombre { get; set; }
        
-        public double Precio_Venta { get; set; }
+        public decimal Precio_Venta { get; set; }
         
         public int Stock { get; set; }
         public string Descripcion { get; set; }
